+++ eras/shelley-ma/test-suite/test/Test/Cardano/Ledger/ShelleyMA/Serialisation/Golden/Encoding.hs
@@ -49,7 +49,7 @@ import qualified Data.Map.Strict as Map
 import qualified Data.Sequence.Strict as StrictSeq
 import qualified Data.Set as Set
 import Lens.Micro
-import Test.Cardano.Ledger.Binary.RoundTrip (roundTripFailureCborRangeExpectation)
+import Test.Cardano.Ledger.Binary.RoundTrip (roundTripCborRangeFailureExpectation)
 import Test.Cardano.Ledger.Shelley.Generator.EraGen (genesisId)
 import Test.Cardano.Ledger.Shelley.Serialisation.GoldenUtils (
   ToTokens (..),
@@ -442,7 +442,7 @@ goldenEncodingTestsMary =
 
 assetName32Bytes :: Assertion
 assetName32Bytes =
-  roundTripFailureCborRangeExpectation (eraProtVerHigh @Mary) maxBound $
+  roundTripCborRangeFailureExpectation (eraProtVerHigh @Mary) maxBound $
     AssetName "123456789-123456789-123456789-123"
 
 -- | Golden Tests for Allegra and Mary