+++ libs/cardano-ledger-binary/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Version history for `cardano-ledger-binary`
 
-## 1.0.1.0
+## 1.1.0.0
 
 * Add `ToJSON`/`FromJSON` instances for `Version`
 * Add `decodeFullFromHexText` and `serializeAsHexText`
@@ -14,6 +14,14 @@
 ### `testlib`
 
 * Add `Arbitrary` instance for `Term`
+* Renamed:
+  * `roundTripAnnFailureRangeExpectation` -> `roundTripAnnRangeFailureExpectation`
+  * `roundTripFailureCborRangeExpectation` -> `roundTripCborRangeFailureExpectation`
+  * `roundTripAnnFailureRangeExpectation` -> `roundTripAnnRangeFailureExpectation`
+* Added:
+  * `embedTripFailureExpectation`
+  * `embedTripRangeFailureExpectation`
+  * `roundTripRangeFailureExpectation`
 
 ## 1.0.0.0